+++ resdk/__init__.py
@@ -1,5 +1,5 @@
 """Resolwe SDK for Python."""
 from __future__ import absolute_import, division, print_function, unicode_literals
 
-from .resolwe import Resolwe, ResolweQuerry
+from .resolwe import Resolwe, ResolweQuery
 from .resdk_logger import log_to_stdout, log_to_file, start_logging