+++ resdk/tests/unit/test_resolwe_query.py
@@ -10,20 +10,20 @@ import slumber
 
 from mock import patch, MagicMock
 
-from resdk.resolwe import ResolweQuerry
+from resdk.resolwe import ResolweQuery
 
 
-class TestResolweQuerry(unittest.TestCase):
+class TestResolweQuery(unittest.TestCase):
 
     @patch('resdk.resolwe.getattr')
     @patch('resdk.resolwe.Resolwe')
     @patch('resdk.resources.sample.Sample')
-    @patch('resdk.resolwe.ResolweQuerry', spec=True)
+    @patch('resdk.resolwe.ResolweQuery', spec=True)
     def test_init(self, resq_mock, resolwe_mock, resource_mock, getattr_mock):
         getattr_mock.return_value = "blah"
-        ResolweQuerry.__init__(resq_mock, resolwe_mock, resource_mock)
+        ResolweQuery.__init__(resq_mock, resolwe_mock, resource_mock)
 
-    @patch('resdk.resolwe.ResolweQuerry', spec=True)
+    @patch('resdk.resolwe.ResolweQuery', spec=True)
     def test_get(self, resq_mock):
 
         # User provides correct ID
@@ -33,12 +33,12 @@ class TestResolweQuerry(unittest.TestCase):
         resq_mock.resource = MagicMock(return_value="Some response")
 
         resq_mock.endpoint = 'anything'
-        data = ResolweQuerry.get(resq_mock, 40)
+        data = ResolweQuery.get(resq_mock, 40)
         self.assertEqual(data, "Some response")
         resq_mock.resource.assert_called_with(id=40, resolwe=resq_mock.resolwe)
 
         resq_mock.endpoint = 'presample'
-        data = ResolweQuerry.get(resq_mock, 40)
+        data = ResolweQuery.get(resq_mock, 40)
         self.assertEqual(data, "Some response")
         resq_mock.resource.assert_called_with(id=40, resolwe=resq_mock.resolwe, presample=True)
 
@@ -46,7 +46,7 @@ class TestResolweQuerry(unittest.TestCase):
         message = r'Id: .* does not exist or you dont have access permission.'
         resq_mock.resource = MagicMock(side_effect=[slumber.exceptions.HttpNotFoundError(message)])
         with six.assertRaisesRegex(self, slumber.exceptions.HttpNotFoundError, message):
-            ResolweQuerry.get(resq_mock, 12345)
+            ResolweQuery.get(resq_mock, 12345)
 
         # User provides correct slug.
         resq_mock.api = MagicMock()
@@ -54,12 +54,12 @@ class TestResolweQuerry(unittest.TestCase):
         resq_mock.resource = MagicMock(return_value="Some response")
 
         resq_mock.endpoint = 'anything'
-        data = ResolweQuerry.get(resq_mock, "abc")
+        data = ResolweQuery.get(resq_mock, "abc")
         self.assertEqual(data, "Some response")
         resq_mock.resource.assert_called_with(slug='abc', resolwe=resq_mock.resolwe)
 
         resq_mock.endpoint = 'presample'
-        data = ResolweQuerry.get(resq_mock, "abc")
+        data = ResolweQuery.get(resq_mock, "abc")
         self.assertEqual(data, "Some response")
         resq_mock.resource.assert_called_with(slug='abc',
                                               resolwe=resq_mock.resolwe, presample=True)
@@ -68,9 +68,9 @@ class TestResolweQuerry(unittest.TestCase):
         message = r'Slug: .* does not exist or you dont have access permission.'
         resq_mock.resource = MagicMock(side_effect=[IndexError(message)])
         with six.assertRaisesRegex(self, IndexError, message):
-            ResolweQuerry.get(resq_mock, "some-slug")
+            ResolweQuery.get(resq_mock, "some-slug")
 
-    @patch('resdk.resolwe.ResolweQuerry', spec=True)
+    @patch('resdk.resolwe.ResolweQuery', spec=True)
     def test_filter(self, resq_mock):
 
         resq_mock.api = MagicMock()
@@ -79,20 +79,20 @@ class TestResolweQuerry(unittest.TestCase):
         resq_mock.resource = MagicMock(return_value=12345)
         resq_mock.endpoint = 'anything_but_presample'
 
-        data = ResolweQuerry.filter(resq_mock, slug="some-slug")
+        data = ResolweQuery.filter(resq_mock, slug="some-slug")
         self.assertIsInstance(data, list)
         self.assertEqual(data[0], 12345)
 
         resq_mock.endpoint = 'presample'
 
-        data = ResolweQuerry.filter(resq_mock, slug="some-slug")
+        data = ResolweQuery.filter(resq_mock, slug="some-slug")
         self.assertIsInstance(data, list)
         self.assertEqual(data[0], 12345)
 
-    @patch('resdk.resolwe.ResolweQuerry', spec=True)
+    @patch('resdk.resolwe.ResolweQuery', spec=True)
     def test_search(self, resq_mock):
         with six.assertRaisesRegex(self, NotImplementedError, ""):
-            ResolweQuerry.search(resq_mock)
+            ResolweQuery.search(resq_mock)
 
 if __name__ == '__main__':
     unittest.main()